misc/HtmlTestSuite/results/test-0185.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Lists: parsing counter-set with valid values</title>
<link href="https://drafts.csswg.org/css-lists-3/#propdef-counter-set" rel="help"/>
<meta content="counter-set supports the full grammar '[ &lt;counter-name&gt; &lt;integer&gt;? ]+ | none'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_valid_value('counter-set', 'none');
test_valid_value('counter-set', 'chapter', 'chapter 0');
test_valid_value('counter-set', 'section -1');
test_valid_value('counter-set', 'first -1 second third 99', 'first -1 second 0 third 99');
</script>
</body>
</html>
